Change the infrastructure & dockerize frontend
We made a twiste to the professor repo by taking the infrastructure out of
symfony app, and change the nginx configuration  fgrom pointing to the
index.php within symfony application into react app

Also, we organized the folder structure by separating each project depending
on its job in the app.

We decided to use React to build the frotend to consume the api that we will
build in symfony and the opther technologies.